pyproject.toml
[tool.poetry]
name = "adventofcode"
version = "0.1.0"
description = ""
authors = ["Christopher Carroll <chris@carroll.codes>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
numpy = ">=1.23.5"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.0"
pytest-watch = ">=4.2.0"
flake8 = ">=6.0.0"

[tool.poetry.group.dagger.dependencies]
dagger-io = ">=0.2.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
